backend/render/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.7.4
requests==2.32.3
pandas==2.2.3
apscheduler==3.10.4
python-dotenv==1.0.1
